# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.ActionType:
    enum:
    - create
    - read
    - update
    - delete
    - login
    - logout
    type: string
    x-enum-varnames:
    - ActionCreate
    - ActionRead
    - ActionUpdate
    - ActionDelete
    - ActionLogin
    - ActionLogout
...
  entity.AuditLog:
    properties:
      action:
        $ref: '#/definitions/entity.ActionType'
      created_at:
        type: string
      detail:
        type: string
      id:
        type: integer
      ip:
        type: string
      resource:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_agent:
        type: string
      user_id:
        type: integer
    type: object
...
  entity.Permission:
    enum:
    - user:create
    - user:read
    - user:update
    - user:delete
    - role:create
    - role:read
    - role:update
    - role:delete
    - audit:read
    - module:integrate
    - warehouse:create
    - warehouse:read
    - warehouse:update
    - warehouse:delete
    - inventory:read
    - inventory:update
    - stock:create
    - stock:read
    - supplier:create
    - supplier:read
    - supplier:update
    - supplier:delete
    - product:create
    - product:read
    - product:update
    - product:delete
    - contract:create
    - contract:read
    - contract:update
    - contract:delete
    - rating:create
    - rating:read
    - manufacturing:facility:create
    - manufacturing:facility:read
    - manufacturing:facility:update
    - manufacturing:facility:delete
    - manufacturing:order:create
    - manufacturing:order:read
    - manufacturing:order:update
    - manufacturing:order:delete
    - manufacturing:bom:create
    - manufacturing:bom:read
    - manufacturing:bom:update
    - manufacturing:bom:delete
    - purchase:request:create
    - purchase:request:read
    - purchase:request:update
    - purchase:request:delete
    - purchase:request:approve
    - purchase:order:create
    - purchase:order:read
    - purchase:order:update
    - purchase:order:delete
    - purchase:order:approve
    - purchase:receipt:create
    - purchase:receipt:read
    - purchase:receipt:update
    - purchase:payment:create
    - purchase:payment:read
    - purchase:payment:update
    - customer:create
    - customer:read
    - customer:update
    - customer:delete
    - customer:address:create
    - customer:address:read
    - customer:address:update
    - customer:address:delete
    - customer:debt:read
    - customer:debt:update
    - customer:loyalty:read
    - customer:loyalty:update
    - sales:order:create
    - sales:order:read
    - sales:order:update
    - sales:order:delete
    - sales:order:confirm
    - sales:order:cancel
    - delivery:order:create
    - delivery:order:read
    - delivery:order:update
    - delivery:order:process
    - invoice:create
    - invoice:read
    - invoice:update
    - invoice:issue
    - invoice:pay
    - finance:invoice:create
    - finance:invoice:read
    - finance:invoice:update
    - finance:invoice:delete
    - finance:payment:create
    - finance:payment:read
    - finance:payment:update
    - finance:payment:process
    - finance:report:read
    - report:create
    - report:read
    - report:update
    - report:delete
    - report:export
    - report:schedule:create
    - report:schedule:read
    - report:schedule:update
    - report:schedule:delete
    type: string
    x-enum-varnames:
    - UserCreate
    - UserRead
    - UserUpdate
    - UserDelete
    - RoleCreate
    - RoleRead
    - RoleUpdate
    - RoleDelete
    - AuditLogRead
    - ModuleIntegrate
    - WarehouseCreate
    - WarehouseRead
    - WarehouseUpdate
    - WarehouseDelete
    - InventoryRead
    - InventoryUpdate
    - StockEntryCreate
    - StockEntryRead
    - SupplierCreate
    - SupplierRead
    - SupplierUpdate
    - SupplierDelete
    - ProductCreate
    - ProductRead
    - ProductUpdate
    - ProductDelete
    - ContractCreate
    - ContractRead
    - ContractUpdate
    - ContractDelete
    - RatingCreate
    - RatingRead
    - ManufacturingFacilityCreate
    - ManufacturingFacilityRead
    - ManufacturingFacilityUpdate
    - ManufacturingFacilityDelete
    - ProductionOrderCreate
    - ProductionOrderRead
    - ProductionOrderUpdate
    - ProductionOrderDelete
    - BOMCreate
    - BOMRead
    - BOMUpdate
    - BOMDelete
    - PurchaseRequestCreate
    - PurchaseRequestRead
    - PurchaseRequestUpdate
    - PurchaseRequestDelete
    - PurchaseRequestApprove
    - PurchaseOrderCreate
    - PurchaseOrderRead
    - PurchaseOrderUpdate
    - PurchaseOrderDelete
    - PurchaseOrderApprove
    - PurchaseReceiptCreate
    - PurchaseReceiptRead
    - PurchaseReceiptUpdate
    - PurchasePaymentCreate
    - PurchasePaymentRead
    - PurchasePaymentUpdate
    - CustomerCreate
    - CustomerRead
    - CustomerUpdate
    - CustomerDelete
    - CustomerAddressCreate
    - CustomerAddressRead
    - CustomerAddressUpdate
    - CustomerAddressDelete
    - CustomerDebtRead
    - CustomerDebtUpdate
    - CustomerLoyaltyRead
    - CustomerLoyaltyUpdate
    - SalesOrderCreate
    - SalesOrderRead
    - SalesOrderUpdate
    - SalesOrderDelete
    - SalesOrderConfirm
    - SalesOrderCancel
    - DeliveryOrderCreate
    - DeliveryOrderRead
    - DeliveryOrderUpdate
    - DeliveryOrderProcess
    - InvoiceCreate
    - InvoiceRead
    - InvoiceUpdate
    - InvoiceIssue
    - InvoicePay
    - FinanceInvoiceCreate
    - FinanceInvoiceRead
    - FinanceInvoiceUpdate
    - FinanceInvoiceDelete
    - FinancePaymentCreate
    - FinancePaymentRead
    - FinancePaymentUpdate
    - FinancePaymentProcess
    - FinanceReportRead
    - ReportCreate
    - ReportRead
    - ReportUpdate
    - ReportDelete
    - ReportExport
    - ReportScheduleCreate
    - ReportScheduleRead
    - ReportScheduleUpdate
    - ReportScheduleDelete
...
  entity.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/entity.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login:
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      role_id:
        type: integer
      status:
        $ref: '#/definitions/entity.UserStatus'
      updated_at:
        type: string
      username:
        type: string
    type: object
  entity.UserStatus:
    enum:
    - active
    - inactive
    - locked
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusLocked
...
  server.AuditLogResponse:
    properties:
      logs:
        items:
          $ref: '#/definitions/entity.AuditLog'
        type: array
      total:
        example: 100
        type: integer
    type: object
...
  server.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/lugondev/erp-warehouse-simple
  description: Simple ERP Warehouse Management System API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ERP Warehouse API
  version: "1.0"
paths:
...
  /api/v1/audit/logs:
    get:
      description: Get a paginated list of all audit logs
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.AuditLogResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all audit logs
      tags:
      - audit
  /api/v1/audit/logs/user/{id}:
    get:
      description: Get audit logs for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.AuditLog'
            type: array
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user audit logs
      tags:
      - audit
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"